[Feature/charger] 충전소 api 수정

See merge request cloud_track/class_01/web_project3/team04/pikacharge_be!21

+++ src/main/java/elice04_pikacharger/pikacharger/domain/charger/service/ChargerServiceImpl.java
@@ -1,8 +1,6 @@
 package elice04_pikacharger.pikacharger.domain.charger.service;
 
-import elice04_pikacharger.pikacharger.domain.charger.dto.ChargerDetailResponseDto;
-import elice04_pikacharger.pikacharger.domain.charger.dto.ChargerEditResponseDto;
-import elice04_pikacharger.pikacharger.domain.charger.dto.ChargerResponseDto;
+import elice04_pikacharger.pikacharger.domain.charger.dto.*;
 import elice04_pikacharger.pikacharger.domain.charger.dto.payload.ChargerCreateDto;
 import elice04_pikacharger.pikacharger.domain.charger.dto.payload.ChargerUpdateDto;
 import elice04_pikacharger.pikacharger.domain.charger.entity.Charger;
@@ -11,6 +9,9 @@ import elice04_pikacharger.pikacharger.domain.charger.repository.ChargerReposito
 import elice04_pikacharger.pikacharger.domain.chargertype.dto.payload.ChargerTypeDto;
 import elice04_pikacharger.pikacharger.domain.chargertype.entity.ChargerType;
 import elice04_pikacharger.pikacharger.domain.favorite.repository.FavoriteRepository;
+import elice04_pikacharger.pikacharger.domain.image.domain.ChargerImage;
+import elice04_pikacharger.pikacharger.domain.image.domain.ReviewImage;
+import elice04_pikacharger.pikacharger.domain.image.service.S3UploaderService;
 import elice04_pikacharger.pikacharger.domain.user.entity.User;
 import elice04_pikacharger.pikacharger.domain.user.repository.UserRepository;
 import jakarta.persistence.EntityNotFoundException;
@@ -18,8 +19,13 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
 
 @RequiredArgsConstructor
 @Slf4j
@@ -30,14 +36,15 @@ public class ChargerServiceImpl implements ChargerService {
     private final UserRepository userRepository;
     private final GeocodingAPI geocodingAPI;
     private final FavoriteRepository favoriteRepository;
+    private final S3UploaderService s3UploaderService;
 
 
     @Transactional
     @Override
-    public ChargerResponseDto createCharger(ChargerCreateDto chargerCreateDto) {
+    public ChargerResponseDto createCharger(ChargerCreateDto chargerCreateDto, List<MultipartFile> multipartFiles) throws IOException{
         User user = userRepository.findById(chargerCreateDto.getUserId())
                 .orElseThrow(() -> new EntityNotFoundException("유저가 존재하지 않습니다."));
-        /** Todo 1.이미지 저장 추가 */
+
         try{
             List<String> locations = geocodingAPI.coordinatePairs(chargerCreateDto.getChargerLocation());
             chargerCreateDto.setLatitude(Double.parseDouble(locations.get(0))); // 위도
@@ -45,6 +52,8 @@ public class ChargerServiceImpl implements ChargerService {
         }catch (Exception e){
             throw new RuntimeException("예외 발생"+e.getMessage());
         }
+
+
 
         Charger charger = chargerCreateDto.toEntity(user);
         for (ChargerTypeDto chargerTypeDto : chargerCreateDto.getChargerTypeDtoList()){
@@ -54,6 +63,33 @@ public class ChargerServiceImpl implements ChargerService {
                     .build();
             charger.getChargerTypes().add(chargerType);
         }
+
+        List<String> imgPaths = new ArrayList<>();
+        if (multipartFiles != null && !multipartFiles.isEmpty()) {
+            imgPaths = s3UploaderService.uploadMultipleFiles(multipartFiles, "images");
+        }
+
+        for (String imgUrl : imgPaths) {
+            ChargerImage image = ChargerImage.builder()
+                    .charger(charger)
+                    .imageUrl(imgUrl)
+                    .build();
+            charger.getChargerImages().add(image);
+        }
+
+//        List<String> imgPaths = new ArrayList<>();
+//        if (chargerCreateDto.getMultipartFiles() != null && !chargerCreateDto.getMultipartFiles().isEmpty()) {
+//            imgPaths = s3UploaderService.uploadMultipleFiles(chargerCreateDto.getMultipartFiles(), "images");
+//        }
+//
+//        for (String imgUrl : imgPaths) {
+//            ChargerImage image = ChargerImage.builder()
+//                    .charger(charger)
+//                    .imageUrl(imgUrl)
+//                    .build();
+//            charger.getChargerImages().add(image);
+//        }
+
         Charger savedCharger = chargerRepository.save(charger);
 
         return ChargerResponseDto.toDto(savedCharger);
@@ -61,7 +97,10 @@ public class ChargerServiceImpl implements ChargerService {
 
     @Transactional
     @Override
-    public ChargerResponseDto updateCharger(ChargerUpdateDto chargerUpdateDto, Long chargerId) {
+    public ChargerResponseDto updateCharger(ChargerUpdateDto chargerUpdateDto, Long chargerId, Long userId) {
+        if (!chargerRepository.existsByIdAndUserId(chargerId, userId)){
+            throw new IllegalStateException("충전소 수정권한이 없습니다.");
+        }
         Charger charger = chargerRepository.findById(chargerId)
                 .orElseThrow(() -> new EntityNotFoundException("충전소가 존재하지 않습니다."));
         try{
@@ -96,7 +135,10 @@ public class ChargerServiceImpl implements ChargerService {
     }
 
     @Override
-    public void deleteCharger(Long chargerId) {
+    public void deleteCharger(Long chargerId, Long userId) {
+        if (!chargerRepository.existsByIdAndUserId(chargerId, userId)){
+            throw new IllegalStateException("충전소 삭제권한이 없습니다.");
+        }
         Charger charger = chargerRepository.findById(chargerId)
                 .orElseThrow(() -> new EntityNotFoundException("충전소가 존재하지 않습니다."));
         chargerRepository.delete(charger);
@@ -112,10 +154,45 @@ public class ChargerServiceImpl implements ChargerService {
     }
 
     @Override
-    public ChargerEditResponseDto chargerEditDetail(Long chargerId) {
+    public ChargerEditResponseDto chargerEditDetail(Long chargerId, Long userId) {
+        if (!chargerRepository.existsByIdAndUserId(chargerId, userId)){
+            throw new IllegalStateException("충전소 수정권한이 없습니다.");
+        }
         Charger charger = chargerRepository.findById(chargerId)
                 .orElseThrow(() -> new EntityNotFoundException("충전소가 존재하지 않습니다."));
 
         return ChargerEditResponseDto.toDto(charger);
+    }
+
+    @Override
+    public List<GroupedChargerResponseDto> chargerSearch(String location) {
+        try{
+            List<String> locations = geocodingAPI.coordinatePairs(location);
+            List<Charger> chargerList = chargerRepository.findChargersNearby(Double.parseDouble(locations.get(0)),Double.parseDouble(locations.get(1)));
+            List<ChargerSearchResponseDto> chargerSearchResponseDtoList = chargerList.stream()
+                    .map(ChargerSearchResponseDto::toDto)
+                    .toList();
+
+            Map<String, List<ChargerSearchResponseDto>> groupedByLocation = chargerSearchResponseDtoList.stream()
+                    .collect(Collectors.groupingBy(dto -> dto.getChargerLocation() + "|" + dto.getChargerName()));
+
+            List<GroupedChargerResponseDto> groupedDtoList = new ArrayList<>();
+            for (var entry : groupedByLocation.entrySet()) {
+                String[] keyParts = entry.getKey().split("\\|", -1);
+                String chargerLocation = keyParts[0];
+                String chargerName = keyParts[1];
+
+                GroupedChargerResponseDto groupedChargerResponseDto = new GroupedChargerResponseDto();
+                groupedChargerResponseDto.setChargerLocation(chargerLocation);
+                groupedChargerResponseDto.setChargerName(chargerName);
+                groupedChargerResponseDto.setChargers(entry.getValue());
+                groupedDtoList.add(groupedChargerResponseDto);
+            }
+
+            return groupedDtoList;
+        }catch (Exception e){
+            log.debug("좌표를 받아올수 없습니다. 오류 코드: "+ e.getMessage());
+        }
+        return null;
     }
 }
